Add home page detail popup, fix list metadata, expose user lists in Stremio addon, add list delete
- Clicking any card on the dashboard (Continue Watching, Recently Watched, Trending, Recommendations) now opens the same detail panel as the search page, showing description, rating, genres, personal star rating, and watch history
- Extract DetailPanel/StarRating into shared MediaDetailPanel component used by both Search and Dashboard pages
- Fix metadata (posters, descriptions) not populating for list items by triggering a background TMDB sync when an item is added to a list
- Expose user-created lists as Stremio catalogs: each enabled list adds separate Movies and Series catalogs to the Stremio manifest, matching the existing Watchlist pattern
- Add user lists section to the Stremio Addon settings so lists can be toggled on/off
- Add delete list button to the Lists page sidebar with inline confirmation
